chore: generation of image and text assets on their own
